# server/openapi.yaml
openapi: 3.0.3
info:
  title: Cook Africa API
  description: |
    API backend de l'application de gestion interne du restaurant Cook Africa,
    plus le point d'entrée public utilisé par la page client `/commander`.

    ### Authentification
    La quasi-totalité des routes exigent un JWT Bearer obtenu via `POST /api/auth/login`,
    envoyé dans l'en-tête `Authorization: Bearer <token>`. Le token expire après 12h.

    Les routes sous `/api/public/*` sont volontairement **non authentifiées** : elles sont
    appelées depuis le navigateur d'un client final (page de commande en ligne), pas depuis
    un poste staff. Elles n'exposent que le menu disponible et la création/consultation de
    commandes ; toute autre donnée reste protégée.

    ### Rôles
    `admin`, `caissiere`, `serveur`, `cuisiniere`, `barman`, `caissier-en-ligne` — la plupart des routes
    restreignent l'accès à un sous-ensemble de rôles via un contrôle serveur (403 sinon).

    ### Restriction Wi-Fi
    Pour tout rôle autre que `admin`, les requêtes authentifiées sont en plus filtrées par
    IP si la restriction Wi-Fi est activée (`/api/wifi-config`) — 403 `wifi_restricted`
    si le poste n'est pas sur le réseau autorisé.

    ### Erreurs
    Toutes les erreurs suivent le même format `{ "error": "message" }`, avec le code HTTP
    approprié (400 validation, 401 non authentifié, 403 accès refusé, 404 introuvable,
    409 conflit, 429 limite de requêtes, 500 erreur serveur).
  version: 1.0.0
  contact:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/commandes:
    get:
      operationId: getApiCommandes
      tags: [Commandes]
      summary: Liste des commandes (staff)
      parameters:
        - name: statut
          in: query
          schema: { $ref: '#/components/schemas/StatutCommande' }
        - name: date
          in: query
          description: Filtre sur la date (YYYY-MM-DD)
          schema: { type: string, format: date }
      responses:
        '200':
          description: Liste des commandes (200 plus récentes, mises en cache 15s)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Commande' }
    post:
      operationId: postApiCommandes
      tags: [Commandes]
      summary: Créer une commande (serveur/caissière/admin)
      description: |
        Un serveur crée une commande au statut `en-attente` (doit ensuite l'envoyer via
        `PUT /{id}/envoyer`). Un admin/caissière crée directement au statut
        `en-preparation`. Seuls admin/caissière peuvent poser `source: "en-ligne"`
        (écran dédié "Commandes en Ligne" côté caisse).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CommandeCreateRequest' }
      responses:
        '201':
          description: Commande créée
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Commande' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/commandes/{id}/envoyer:
    put:
      operationId: putApiCommandesByIdEnvoyer
      tags: [Commandes]
      summary: Envoyer une commande à la facturation (serveur/admin)
      description: Passe la commande de `en-attente` à `en-preparation` — elle appartient ensuite à la caissière.
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Commande envoyée
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Commande' }
        '400':
          description: Déjà envoyée à la facturation
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/commandes/{id}/items:
    put:
      operationId: putApiCommandesByIdItems
      tags: [Commandes]
      summary: Modifier librement les articles d'une commande
      description: Interdit si la commande est terminée, si elle a déjà une facture de type "facture", ou si un serveur tente de modifier une commande déjà envoyée en caisse.
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  items: { $ref: '#/components/schemas/CommandeItemInput' }
      responses:
        '200':
          description: Commande modifiée
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Commande' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/commandes/{id}:
    put:
      operationId: putApiCommandesById
      tags: [Commandes]
      summary: Mettre à jour le statut, la note ou le n° de table d'une commande
      description: Seuls `statut`, `note` et `tableNumero` sont modifiables via cette route (whitelist anti mass-assignment) ; `total`/`items`/`createdBy` sont ignorés silencieusement.
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CommandeUpdateRequest' }
      responses:
        '200':
          description: Commande mise à jour
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Commande' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteApiCommandesById
      tags: [Commandes]
      summary: Annuler une commande (admin)
      description: Passe la commande au statut `annulee` et supprime la facture liée si elle existe.
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Commande annulée
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  factureSupprimee: { type: boolean }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ── Commande publique (client) ───────────────────────
...
  /api/public/commandes:
    post:
      operationId: postApiPublicCommandes
      tags: [Commande publique]
      summary: Le client passe sa propre commande depuis son téléphone
      description: |
        Aucune authentification. Limité à 10 requêtes / 10 min par IP.
        **Le prix et le nom de chaque article sont toujours résolus côté serveur à
        partir du menu réel** — jamais depuis ce que le client envoie — pour empêcher
        toute falsification de prix. La commande part directement au statut
        `en-preparation` avec `source: "en-ligne"` et `commandeClient: true`, et
        apparaît immédiatement dans l'onglet "Commandes en Ligne" côté caisse.
        Prénom, nom, téléphone et localisation sont obligatoires (livraison) ;
        `localisation.mapsUrl` est toujours reconstruit côté serveur à partir de
        `lat`/`lng`, jamais accepté tel quel depuis le client.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PublicCommandeCreateRequest' }
      responses:
        '201':
          description: Commande créée
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicCommandeCreateResponse' }
        '400':
          description: Panier vide, article introuvable/indisponible, quantité invalide, ou trop d'articles
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429':
          description: Trop de commandes envoyées depuis cette IP
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/public/commandes/{id}:
    get:
      operationId: getApiPublicCommandesById
      tags: [Commande publique]
      summary: Suivi de statut d'une commande client
      description: |
        L'`id` Firestore renvoyé à la création sert de jeton d'accès (non énumérable).
        Ne renvoie que les commandes marquées `commandeClient: true` — une commande
        staff n'est jamais consultable par cette route.
      security: []
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Statut courant de la commande
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicCommandeStatus' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ── Menu ──────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Jeton obtenu via `POST /api/auth/login`, valable 12h.
  parameters:
    IdPathParam:
      name: id
      in: path
      required: true
      description: Identifiant du document Firestore
      schema: { type: string }
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Forbidden:
      description: Accès refusé (rôle insuffisant, ou hors réseau Wi-Fi autorisé)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    StatutCommande:
      type: string
      enum: [en-attente, en-preparation, prete, servie, annulee]
    SourceCommande:
      type: string
      enum: [sur-place, en-ligne]
...
    CommandeItem:
      type: object
      properties:
        menuItemId: { type: string }
        nom: { type: string }
        prix: { type: number }
        quantite: { type: integer }
        sousTotal: { type: number }
        categorie: { type: string }
    CommandeItemInput:
      type: object
      required: [nom, prix, quantite]
      properties:
        menuItemId: { type: string }
        nom: { type: string }
        prix: { type: number }
        quantite: { type: integer, minimum: 1 }
        categorie: { type: string }
    Commande:
      type: object
      properties:
        id: { type: string }
        numero: { type: string, example: CMD-0001 }
        items:
          type: array
          items: { $ref: '#/components/schemas/CommandeItem' }
        total: { type: number }
        note: { type: string }
        tableNumero: { type: string }
        source: { $ref: '#/components/schemas/SourceCommande' }
        commandeClient:
          type: boolean
          description: Présent et vrai si la commande a été passée par le client lui-même via /commander.
        clientPrenom: { type: string, description: Renseigné uniquement pour une commande client. }
        clientNom: { type: string, description: Renseigné uniquement pour une commande client. }
        clientTelephone: { type: string, description: Renseigné uniquement pour une commande client. }
        clientLocalisation:
          $ref: '#/components/schemas/Localisation'
        statut: { $ref: '#/components/schemas/StatutCommande' }
        date: { type: string, format: date }
        createdBy: { type: string }
        createdByNom: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        envoyeeAt: { type: string, format: date-time }
        envoyeeBy: { type: string }
        envoyeeByNom: { type: string }
        lastEditedBy: { type: string }
        lastEditedByNom: { type: string }
        lastEditedAt: { type: string, format: date-time }
    CommandeCreateRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/CommandeItemInput' }
        note: { type: string }
        tableNumero: { type: string }
        source:
          type: string
          enum: [en-ligne]
          description: Optionnel — seuls admin/caissière peuvent l'utiliser pour marquer la commande "en ligne".
    CommandeUpdateRequest:
      type: object
      description: Whitelist stricte — tout autre champ envoyé est ignoré.
      properties:
        statut: { $ref: '#/components/schemas/StatutCommande' }
        note: { type: string }
        tableNumero: { type: string }
    PublicCommandeItemInput:
      type: object
      required: [menuItemId, quantite]
      properties:
        menuItemId: { type: string }
        quantite: { type: integer, minimum: 1, maximum: 20 }
    Localisation:
      type: object
      description: Reconstruite côté serveur à partir de lat/lng validés — mapsUrl n'est jamais accepté depuis le client.
      properties:
        lat: { type: number, format: double, example: 5.359951 }
        lng: { type: number, format: double, example: -4.008256 }
        mapsUrl: { type: string, format: uri, example: 'https://www.google.com/maps?q=5.359951,-4.008256' }
    PublicCommandeCreateRequest:
      type: object
      required: [items, prenom, nom, telephone, localisation]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 30
          items: { $ref: '#/components/schemas/PublicCommandeItemInput' }
        prenom: { type: string, maxLength: 60 }
        nom: { type: string, maxLength: 60 }
        telephone: { type: string, maxLength: 30, description: '8 à 15 chiffres une fois les caractères non numériques retirés.' }
        localisation:
          type: object
          required: [lat, lng]
          description: Coordonnées GPS capturées via l'API de géolocalisation du navigateur du client.
          properties:
            lat: { type: number, format: double, minimum: -90, maximum: 90 }
            lng: { type: number, format: double, minimum: -180, maximum: 180 }
    PublicCommandeCreateResponse:
      type: object
      properties:
        id: { type: string }
        numero: { type: string }
        items:
          type: array
          items: { $ref: '#/components/schemas/CommandeItem' }
        total: { type: number }
        statut: { $ref: '#/components/schemas/StatutCommande' }
        createdAt: { type: string, format: date-time }
    PublicCommandeStatus:
      type: object
      properties:
        id: { type: string }
        numero: { type: string }
        statut: { $ref: '#/components/schemas/StatutCommande' }
        items:
          type: array
          items: { $ref: '#/components/schemas/CommandeItem' }
        total: { type: number }
        createdAt: { type: string, format: date-time }
...